btstack_uart_block_posix: inline btstack_uart_posix
