+++ platform/posix/hci_transport_h5_posix.c
@@ -53,7 +53,6 @@
 #include "btstack_slip.h"
 #include "btstack_debug.h"
 #include "hci_transport.h"
-#include "btstack_uart_posix.h"
 #include "btstack_uart_block.h"
 
 #ifdef HAVE_EHCILL
