+++ port/daemon/src/Makefile.in
@@ -39,7 +39,6 @@ libBTstack_SOURCES =    \
 			  
 BTdaemon_SOURCES =      \
     daemon.c            \
-    btstack_uart_posix.c \
     btstack_uart_block_posix.c \
     hci_transport_h4_posix.c  \
     $(libBTstack_SOURCES) \
